removed crazy template thing from TouchButtonManager class split TouchButtonManager into separate .cpp and .h files
